Refactoring for upcoming platform independent solution.
* Introduces new factory mathods for PsBasedProcessTree and PsProcess.
* PsBasedProcessTree class is abstract
* PsProcess class is abstract and extracted from PsBasedProcessTree class (was nested before)
* NEW Prototype classes for windows
* NEW classes for unix specific implementation
* NEW PsProcessWin and PsProcessUnix because of kill method specifics on different platform.
