+++ src/main/java/org/jenkinsci/plugins/proccleaner/PsBasedWinProcessTree.java
@@ -1,0 +1,17 @@
+package org.jenkinsci.plugins.proccleaner;
+
+import java.io.IOException;
+
+/**
+ * Author: psrna
+ * Date: 8/2/13
+ */
+public class PsBasedWinProcessTree extends PsBasedProcessTree {
+
+    @Override
+    public PsBasedProcessTree createProcessTreeFor(String user) throws InterruptedException, IOException {
+        return null;  //To change body of implemented methods use File | Settings | File Templates.
+    }
+
+
+}
