+++ src/main/java/org/jenkinsci/plugins/proccleaner/PsKiller.java
@@ -12,13 +12,13 @@ import jenkins.model.Jenkins;
 public abstract class PsKiller implements ExtensionPoint, Serializable {
 
 	public void kill(String user) throws InterruptedException, IOException {
-		PsBasedProcessTree ptree = PsBasedProcessTree.createProcessTreeFor(user);
+		PsBasedProcessTree ptree = PsBasedProcessTreeFactory.createPsBasedProcessTree().createProcessTreeFor(user);
 		int me = ProcCleaner.getpid();
 		doKill(ptree, me);
 	}
 	
 	public void kill(String user, PrintStream log) throws InterruptedException, IOException {
-		PsBasedProcessTree ptree = PsBasedProcessTree.createProcessTreeFor(user);
+		PsBasedProcessTree ptree = PsBasedProcessTreeFactory.createPsBasedProcessTree().createProcessTreeFor(user);
 		System.out.println("Process tree: " + ptree.toString());
 		ptree.setLog(log);
 		int me = ProcCleaner.getpid();
